Criado serviço de log

+++ internal/domain/log/log.go
@@ -1,8 +1,8 @@
 package log
 
 type Log struct {
-	ID      int    `json:"id"`
-	UserID  string `json:"user_id"`
-	EventID int    `json:"event_id"`
-	Score   int    `json:"score"`
+	ID      int     `json:"id"`
+	UserID  string  `json:"user_id"`
+	EventID int     `json:"event_id"`
+	Score   float32 `json:"score"`
 }
